backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
python-dotenv>=1.0.0
anthropic>=0.49.0
supabase>=2.4.0
python-multipart>=0.0.9
pdfplumber>=0.10.0
reportlab>=4.0.0
